Week13/hh_back/requirements.txt
djangorestframework-jwt
appdirs==1.4.3
asgiref==3.2.5
distlib==0.3.0
Django==3.0.4
djangorestframework==3.11.0
filelock==3.0.12
importlib-metadata==1.5.0
importlib-resources==1.3.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
virtualenv==20.0.10
zipp==3.1.0
